+++ src/evict/commands/default_status.rs
@@ -19,7 +19,7 @@
 
 use status_storage;
 
-pub fn default_status(mut args:Vec<String>) -> int {
+pub fn default_status(mut args:Vec<String>) -> isize {
   if args.len() > 1 {
     println! ("default-status usage: evict default-status [new-status]");
     1
